+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Change log
 
-## [XXXXX] - 2024-XXX-XX
+## [1.0.7] - 2024-XXX-XX
+
+Added:
+
+- Niche cluster connectivity visualization for each sample
+- `scale factor` controlling the size of spatial-based plots
 
 Changed:
 
-- `log` file is optional for `ONTraC_analysis`
+- Make `log` file optional for `ONTraC_analysis`
 
 Fixed:
 
@@ -208,5 +213,5 @@ Changed:
 - Running process
 - Uniform parameters control
 - Output directories
-- `createDataSet` generate cell type compostion and GNN input from raw input
+- `createDataSet` generate cell type composition and GNN input from raw input
 - Input data format